requirements.txt
aiofiles==23.2.1
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.27.0
attrs==23.2.0
bcrypt==4.1.2
black==23.12.1
boto3==1.21.31
botocore==1.24.46
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
comm==0.2.1
cryptography==41.0.7
debugpy==1.8.1
decorator==5.1.1
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0
executing==2.0.1
fastapi==0.104.1
flake8==7.0.0
flake8-black==0.3.6
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
ipykernel==6.29.2
ipython==8.21.0
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jwt==1.3.1
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.6.0
# numpy==1.26.3
packaging==23.2
pandas==2.1.4
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.1.0
pluggy==1.3.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==15.0.2
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.2.0
Pygments==2.17.2
pytest==7.1.2
pytest-asyncio==0.23.6
pytest-dotenv==0.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.5
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
requests-toolbelt==0.9.1
rsa==4.9
s3transfer==0.5.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.36
stack-data==0.6.3
starlette==0.27.0
tomli==2.0.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
uvicorn==0.17.6
uvloop==0.19.0
watchgod==0.8.2
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
isort==6.0.0

# fastapi==0.110.0
modal==0.62
# pandas~=1.5.0
# pydantic~=1.10.0
# pytest~=8.0.0
# requests~=2.31.0
numpy==1.26.4
# sqlalchemy~=1.4.0
# psycopg2~=2.9.0
# python-dotenv
nltk==3.6.0
